Change "secondCounter" type, so that it doesn't easily overflow.

+++ aiko_arduinolab.pde
@@ -69,7 +69,7 @@ void loop(void) {
 
 /* ------------------------------------------------------------------------- */
 
-int secondCounter = 0;
+long secondCounter = 0;
 
 void clockHandler(void) {
   ++ secondCounter;
